+++ packages/react-esc-webpack/src/middlewares/index.js
@@ -1,3 +1,1 @@
-export { default as devMiddleware } from './DevMiddleware'
-export { default as hmrMiddleware } from './HmrMiddleware'
 export { default as universalMiddleware } from './UniversalMiddleware'